Extract ActionsAlertDialogCreator
